fix(user/follow-unfollow): added constraint, exception for user can not follow or unfollow himself/herself
added database constraint
added exception throw
added exception

+++ app/Http/Controllers/UserController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\UserFollowException;
 use App\Models\User;
 use App\Repositories\UserRepository;
 use Illuminate\Http\JsonResponse;
@@ -35,6 +36,11 @@ class UserController extends Controller
     public function follow(User $user): JsonResponse
     {
         $authUserId = auth()->user()->id;
+
+        if ($user->id === $authUserId) {
+            throw new UserFollowException('user can not follow themselves');
+        }
+
         $this->userRepository->addFollower($user, $authUserId);
 
         return response()->json(['message' => 'success'], Response::HTTP_ACCEPTED);
@@ -43,6 +49,11 @@ class UserController extends Controller
     public function unfollow(User $user): JsonResponse
     {
         $authUserId = auth()->user()->id;
+
+        if ($user->id === $authUserId) {
+            throw new UserFollowException('user can not unfollow themselves');
+        }
+
         $this->userRepository->removeFollower($user, $authUserId);
 
         return response()->json(['message' => 'success'], Response::HTTP_ACCEPTED);
